requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
yt-dlp
python-dotenv==1.0.0
pyngrok==7.0.5
websockets==12.0
